Extract more of the Sorbet runtime loading into the separate file
By extracting all Sorbet runtime loading into the separate file, we can more easily make changes to how we load Sorbet runtime when we are loading it for executable entrypoints of Ruby LSP.

This commit also drops the `TypeVariableSyntax` patch which is no longer needed, and adds a patch for `T.unsafe`.
