+++ exe/ruby-lsp-check
@@ -3,17 +3,7 @@
 
 # This executable checks if all automatic LSP requests run successfully on every Ruby file under the current directory
 
-require "sorbet-runtime"
-require "ruby_lsp/disable_sorbet"
-
-begin
-  T::Configuration.default_checked_level = :never
-  T::Configuration.call_validation_error_handler = ->(*) {}
-  T::Configuration.inline_type_error_handler = ->(*) {}
-  T::Configuration.sig_validation_error_handler = ->(*) {}
-rescue
-  nil
-end
+require "ruby_lsp/load_sorbet"
 
 $LOAD_PATH.unshift(File.expand_path("../lib", __dir__))
 require "ruby_lsp/internal"
